+++ adapter/mintegral/src/main/java/org/bidon/mintegral/MintegralAdapter.kt
@@ -7,6 +7,7 @@ import com.mbridge.msdk.mbbid.out.BidManager
 import com.mbridge.msdk.out.MBridgeSDKFactory
 import com.mbridge.msdk.out.SDKInitStatusListener
 import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.suspendCancellableCoroutine
 import kotlinx.coroutines.withContext
 import org.bidon.mintegral.ext.adapterVersion
 import org.bidon.mintegral.ext.sdkVersion
@@ -28,7 +29,6 @@ import org.json.JSONObject
 import java.util.concurrent.atomic.AtomicBoolean
 import kotlin.coroutines.resume
 import kotlin.coroutines.resumeWithException
-import kotlin.coroutines.suspendCoroutine
 
 /**
  * Created by Aleksei Cherniaev on 20/06/2023.
@@ -63,7 +63,7 @@ internal class MintegralAdapter :
         if (isInitialized.get()) return
 
         withContext(Dispatchers.Main.immediate) {
-            suspendCoroutine { continuation ->
+            suspendCancellableCoroutine { continuation ->
                 this@MintegralAdapter.context = context
                 val sdk = MBridgeSDKFactory.getMBridgeSDK()
                 val configurationMap = sdk.getMBConfigurationMap(configParams.appId, configParams.appKey)
@@ -72,15 +72,15 @@ internal class MintegralAdapter :
                     context.applicationContext as Application,
                     object : SDKInitStatusListener {
                         override fun onInitSuccess() {
-                            if (isInitialized.compareAndSet(false, true)) {
+                            if (continuation.isActive) {
                                 continuation.resume(Unit)
                             }
                         }
 
                         override fun onInitFail(message: String?) {
-                            if (isInitialized.compareAndSet(false, true)) {
-                                val error = BidonError.Unspecified(demandId, Throwable(message))
-                                logError(TAG, "Error while initialization: $message", error)
+                            val error = BidonError.Unspecified(demandId, Throwable(message))
+                            logError(TAG, "Error while initialization: $message", error)
+                            if (continuation.isActive) {
                                 continuation.resumeWithException(error)
                             }
                         }