+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 ## Features:
 - [BDN-1011](https://appodeal.atlassian.net/browse/BDN-1011) Decoupled adapter version from sdk core version
 - [BDN-1048](https://appodeal.atlassian.net/browse/BDN-1048) Added explicitApi api compile strict mode
+- [BDN-1066](https://appodeal.atlassian.net/browse/BDN-1066) Fixed IllegalStateException: Already resumed in Mintegral adapter
 
 # 0.11.0 (2025.09.10)
 ## Features: